fix(adapters): make testcontainers-go fixture opt-in via dockerfixture build tag (#929)
Gate `adapters/v1/grype_docker_fixture.go` and tests calling `NewGrypeAdapterFixedDB()` (`adapters/v1/grype_docker_fixture_test.go` and `core/services/scan_docker_fixture_test.go`) behind the `//go:build dockerfixture` tag.

This ensures untagged production builds (`go build ./cmd/http`) and dependency checks (`go list -deps ./cmd/http/...`) exclude `testcontainers-go` and `github.com/docker/docker/client` by default without needing any build tags. Update `Makefile`'s `test` target to pass `-tags dockerfixture` for integration tests requiring the real Docker DB container.
